requirements.txt
asgiref==3.8.1
diploma-frontend @ file:///home/alf/PycharmProjects/python_django_diploma/diploma-frontend/dist/diploma-frontend-0.6.tar.gz#sha256=3a6b1203724b256946919b4d65b362b005ae6a9e1e93d8d765e96aa6a92764ac
Django==4.2.13
djangorestframework==3.15.1
pillow==10.3.0
sqlparse==0.5.0
